finalize medianreport codec + split as separate utility
