+++ relayer/pkg/chainlink/ocr2/medianreport/report_fuzz_test.go
@@ -1,7 +1,7 @@
 //go:build go1.18
 // +build go1.18
 
-package ocr2
+package medianreport
 
 import (
 	"math/big"
