+++ relayer/pkg/chainlink/ocr2/medianreport/report_test.go
@@ -1,4 +1,4 @@
-package ocr2
+package medianreport
 
 import (
 	"fmt"
